% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.TimeGPT}
\alias{forecast.TimeGPT}
\title{Forecasting using TimeGPT model}
\usage{
forecast(object, h = 10, level=c(80, 95), xreg,
                token = NULL, verbose = FALSE, ...)
}
\arguments{
\item{object}{An object of class "\code{ets}". Usually the result of a call
to \code{\link{ets}}.}

\item{h}{Number of periods for forecasting}

\item{level}{Confidence level for prediction intervals.}

\item{xreg}{Optionally, a numerical vector or matrix of external regressors, which must have
the same number of rows as y.}

\item{token}{The authorization token to interact with the \code{TimeGPT API}.}

\item{verbose}{If \code{TRUE}, additional console information about the API request and
response are provided.}

\item{...}{Other arguments.}
}
\value{
An object of class "\code{forecast}".
}
\description{
Returns forecasts and other information for the univariate \code{TimeGPT} model.
}
\author{
Daniele Girolimetto
}
