% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeGPT.R
\name{TimeGPT}
\alias{TimeGPT}
\alias{print.TimeGPT}
\title{Fit TimeGPT model to univariate time series}
\usage{
TimeGPT(y, time, freq, xreg, token = NULL, clean_ex_first = TRUE,
               finetune_steps = 0, verbose = FALSE, historic = FALSE)
}
\arguments{
\item{y}{A numeric vector or time series of class \code{ts}.}

\item{time}{A \code{date} vector which must have the same number of rows as \code{y}.}

\item{freq}{A chracter identifying the frequency: \code{D} for daily, \code{M} for monthly,
\code{H} for hourly, and \code{W} for weekly.}

\item{xreg}{Optionally, a numerical vector or matrix of external regressors, which must have
the same number of rows as \code{y}.}

\item{token}{The authorization token to interact with the \code{TimeGPT API}.}

\item{clean_ex_first}{If \code{TRUE}, the exogenous signal is cleaned, otherwise the exogenous
variables are applied after the large time model.}

\item{finetune_steps}{The number of tuning steps used to train the large time model on the data.}

\item{verbose}{If \code{TRUE}, additional console information about the API request and
response are provided.}

\item{historic}{If \code{TRUE}, the fitted time series data for the in-sample period are provided.}
}
\value{
An object of class "\code{TimeGPT}".
}
\description{
Returns an \code{TimeGPT} model applied to y.
}
\author{
Daniele Girolimetto
}
