% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeGPTapi-package.R
\docType{package}
\name{TimeGPTapi-package}
\alias{TimeGPTapi}
\alias{TimeGPTapi-package}
\title{TimeGPTapi: Interact with the 'TimeGPT' API}
\description{
Interface to interact with the 'TimeGPT' API developed by Nixtla \url{https://nixtla.io/}. TimeGPT is a generative pre-trained transformer model specialized in prediction \url{https://docs.nixtla.io/docs}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danigiro/TimeGPTapi}
  \item Report bugs at \url{https://github.com/danigiro/TimeGPTapi/issues}
}

}
\author{
\strong{Maintainer}: Daniele Girolimetto \email{daniele.girolimetto@phd.unipd.it} (\href{https://orcid.org/0000-0001-9387-1232}{ORCID}) [funder]

}
\keyword{internal}
