% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callTimeGPT.R
\name{get_timeGPT_token}
\alias{get_timeGPT_token}
\alias{set_TimeGPT_token}
\title{Set and get the TimeGPT token in the R environment}
\usage{
get_timeGPT_token()

set_TimeGPT_token(token)
}
\arguments{
\item{token}{The authorization token to interact with the \code{TimeGPT API}
\url{https://dashboard.nixtla.io}.}
}
\value{
\code{get_timeGPT_token} returns the token.
}
\description{
Set the \code{TimeGPT} token as an environment variable. You should add
\code{TIMEGPT_TOKEN = <token>} to your \code{.Renviron} otherwise you will
have to call this function every session.
}
