% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeGPTapi-package.R
\docType{package}
\name{TimeGPTapi-package}
\alias{TimeGPTapi}
\alias{TimeGPTapi-package}
\title{TimeGPTapi: Interact with the TimeGPT API}
\description{
\code{TimeGPT} is a generative pre-trained transformer model for time series
analysis (\url{https://docs.nixtla.io/docs}) developed by  \bold{Nixtla}
(\url{https://nixtla.io}). It was trained on the largest collection of public
time series and can predict the future values of a single time series based
on the provided. This R package provides
an R interface to interact with the \code{TimeGPT API}.
}
\section{Copyright}{

This R package is released under the \emph{GNU General Public License v3.0}.
The \code{TimeGPT} model and \code{API} are proprietary technologies developed
by \bold{Nixtla}, and all rights are reserved. The use of the \code{TimeGPT API}
is subject to Nixtla's terms and conditions, which can be found at \url{https://nixtla.io/}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danigiro/TimeGPTapi}
  \item Report bugs at \url{https://github.com/danigiro/TimeGPTapi/issues}
}

}
\author{
Daniele Girolimetto
}
\keyword{package}
